not change in data stored in component 1 of jwords.ijf
Backup and Restore timestamps have been appended to the pack count in component 1 of jwords.ijf.  This changes the datatype of the stored list. backnum has been adjusted to handle this new layout and extant data. Timestamps are included because with the implementation of bget the actual date a backup was made has become more important.